updating client and crypto libraries

+++ anonymous_tokens/cpp/crypto/anonymous_tokens_pb_openssl_converters.cc
@@ -32,7 +32,8 @@ absl::StatusOr<std::string> GenerateMask(
     const RSABlindSignaturePublicKey& public_key) {
   std::string mask;
   if (public_key.message_mask_type() == AT_MESSAGE_MASK_CONCAT &&
-      public_key.message_mask_size() >= kRsaMessageMaskSizeInBytes32) {
+      static_cast<size_t>(public_key.message_mask_size()) >=
+          kRsaMessageMaskSizeInBytes32) {
     mask = std::string(public_key.message_mask_size(), '\0');
     RAND_bytes(reinterpret_cast<uint8_t*>(mask.data()), mask.size());
   } else if (public_key.message_mask_type() == AT_MESSAGE_MASK_NO_MASK &&
